Move web/ to lib/constable_web for Phx 1.3.0.
Phoenix 1.3.x applications have moved to a directory structure that more
closely resembles other mix projects.

This means that we are moving top level `web/` directory to
`lib/constable_web/`, and all controllers, channels, views, the router,
and the endpoint are all now namespaced by `ConstableWeb` instead of
`Constable`.

We follow the instructions described in
https://gist.github.com/chrismccord/71ab10d433c98b714b75c886eff17357