Update brunch_config to match new directory structure
Since we have moved the top level directory `web` to
`lib/constable_web`, and since the `static` directory resides under
`web`, we update the `brunch_config` to point to `lib/constable_web`.

The end goal is to move all assets under a top-level `assets` directory
as is the case with new phoenix 1.3 apps.