Update mix test to account for assets being moved
We add `bin/test_suite` script to run full test suite after building
assets. We can no longer simply alias the `mix test` command to run
`brunch build` before doing the rest of the usual test stuff because the
brunch config is now under the `assets` directory and not in the root of
the project. Brunch expects the calling directory to have the
`brunch-config.js` file in it. So we cannot simply call
`assets/node_modules/brunch/bin/brunch build` because brunch would
expect the file to be in the root directory instead of in `assets`.

We update the README to reflect the new way to run the entire test
suite via `bin/test_suite`.