Allow <iframe>s in announcements
We change `HtmlSanitizeEx` from `basic_html` to `html5`. That function
is still a work-in-progress according to some comments in the source
code, but I think it is safe enough for us to use since this is an
internal tool.

It would still be great to add oembed so that someone can just put a
link and we can retrieve it for them, but allowing the rendering of
<iframe>s is a pre-requisite for that anyway.